Add files via upload
here in this repository i explained about Linear_algebra for  machine learning ..! 

Linear Algebra is the branch of mathematics that studies  vector spaces and linear transformations between vector spaces, such as rotating a shape, scaling it up or down, translating it (ie. moving it), etc.

Machine Learning relies heavily on Linear Algebra, so it is essential to understand what vectors and matrices are, what operations you can perform with them, and how they can be useful.